@@ -1,36 +1,275 @@
 {
-  "name": "vite-project",
-  "private": true,
-  "version": "0.0.0",
+  "name": "seminar2",
+  "version": "1.0.0",
+  "description": "",
+  "homepage": "https://github.com/Atlaseer/fullstacks2#readme",
+  "bugs": {
+    "url": "https://github.com/Atlaseer/fullstacks2/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Atlaseer/fullstacks2.git"
+  },
+  "license": "ISC",
+  "author": "",
+  "type": "commonjs",
+  "main": "index.js",
   "scripts": {
     "dev": "concurrently \"npm run server\" \"npm run client\"",
-    "server": "nodemon server/index.js",
-    "client": "cd client && npm run dev:client",
-    "dev:client": "vite",
-    "build": "cd client && vite build",
-    "preview": "cd client && vite preview",
-    "lint": "ce client && eslint ."
+    "server": "cd server && npm run dev",
+    "client": "cd client && npm run dev",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "dependencies": {
+    "abbrev": "^1.1.1",
+    "accepts": "^2.0.0",
+    "acorn": "^8.14.1",
+    "acorn-jsx": "^5.3.2",
+    "agent-base": "^6.0.2",
+    "ajv": "^6.12.6",
+    "ansi-regex": "^5.0.1",
+    "ansi-styles": "^4.3.0",
+    "anymatch": "^3.1.3",
+    "aproba": "^2.0.0",
+    "are-we-there-yet": "^2.0.0",
+    "argparse": "^2.0.1",
+    "balanced-match": "^1.0.2",
     "bcrypt": "^5.1.1",
-    "concurrently": "^9.1.2",
+    "binary-extensions": "^2.3.0",
+    "body-parser": "^2.2.0",
+    "brace-expansion": "^1.1.11",
+    "braces": "^3.0.3",
+    "browserslist": "^4.24.4",
+    "bson": "^6.10.3",
+    "bytes": "^3.1.2",
+    "call-bind-apply-helpers": "^1.0.2",
+    "call-bound": "^1.0.4",
+    "callsites": "^3.1.0",
+    "caniuse-lite": "^1.0.30001715",
+    "chalk": "^4.1.2",
+    "chokidar": "^3.6.0",
+    "chownr": "^2.0.0",
+    "cliui": "^8.0.1",
+    "color-convert": "^2.0.1",
+    "color-name": "^1.1.4",
+    "color-support": "^1.1.3",
+    "concat-map": "^0.0.1",
+    "console-control-strings": "^1.1.0",
+    "content-disposition": "^1.0.0",
+    "content-type": "^1.0.5",
+    "convert-source-map": "^2.0.0",
+    "cookie": "^0.7.2",
+    "cookie-signature": "^1.2.2",
     "cors": "^2.8.5",
+    "cross-spawn": "^7.0.6",
+    "csstype": "^3.1.3",
+    "debug": "^4.4.0",
+    "deep-is": "^0.1.4",
+    "delegates": "^1.0.0",
+    "depd": "^2.0.0",
+    "detect-libc": "^2.0.4",
     "dotenv": "^16.5.0",
+    "dunder-proto": "^1.0.1",
+    "ee-first": "^1.1.1",
+    "electron-to-chromium": "^1.5.140",
+    "emoji-regex": "^8.0.0",
+    "encodeurl": "^2.0.0",
+    "es-define-property": "^1.0.1",
+    "es-errors": "^1.3.0",
+    "es-object-atoms": "^1.1.1",
+    "esbuild": "^0.25.3",
+    "escalade": "^3.2.0",
+    "escape-html": "^1.0.3",
+    "escape-string-regexp": "^4.0.0",
+    "eslint": "^9.25.1",
+    "eslint-plugin-react-hooks": "^5.2.0",
+    "eslint-plugin-react-refresh": "^0.4.20",
+    "eslint-scope": "^8.3.0",
+    "eslint-visitor-keys": "^4.2.0",
+    "espree": "^10.3.0",
+    "esquery": "^1.6.0",
+    "esrecurse": "^4.3.0",
+    "estraverse": "^5.3.0",
+    "esutils": "^2.0.3",
+    "etag": "^1.8.1",
     "express": "^5.1.0",
+    "fast-deep-equal": "^3.1.3",
+    "fast-json-stable-stringify": "^2.1.0",
+    "fast-levenshtein": "^2.0.6",
+    "fdir": "^6.4.4",
+    "file-entry-cache": "^8.0.0",
+    "fill-range": "^7.1.1",
+    "finalhandler": "^2.1.0",
+    "find-up": "^5.0.0",
+    "flat-cache": "^4.0.1",
+    "flatted": "^3.3.3",
+    "forwarded": "^0.2.0",
+    "fresh": "^2.0.0",
+    "fs-minipass": "^2.1.0",
+    "fs.realpath": "^1.0.0",
+    "function-bind": "^1.1.2",
+    "gauge": "^3.0.2",
+    "gensync": "^1.0.0-beta.2",
+    "get-caller-file": "^2.0.5",
+    "get-intrinsic": "^1.3.0",
+    "get-proto": "^1.0.1",
+    "glob": "^7.2.3",
+    "glob-parent": "^6.0.2",
+    "globals": "^16.0.0",
+    "gopd": "^1.2.0",
+    "has-flag": "^4.0.0",
+    "has-symbols": "^1.1.0",
+    "has-unicode": "^2.0.1",
+    "hasown": "^2.0.2",
+    "http-errors": "^2.0.0",
+    "https-proxy-agent": "^5.0.1",
+    "iconv-lite": "^0.6.3",
+    "ignore": "^5.3.2",
+    "ignore-by-default": "^1.0.1",
+    "import-fresh": "^3.3.1",
+    "imurmurhash": "^0.1.4",
+    "inflight": "^1.0.6",
+    "inherits": "^2.0.4",
+    "ipaddr.js": "^1.9.1",
+    "is-binary-path": "^2.1.0",
+    "is-extglob": "^2.1.1",
+    "is-fullwidth-code-point": "^3.0.0",
+    "is-glob": "^4.0.3",
+    "is-number": "^7.0.0",
+    "is-promise": "^4.0.0",
+    "isexe": "^2.0.0",
+    "js-tokens": "^4.0.0",
+    "js-yaml": "^4.1.0",
+    "jsesc": "^3.1.0",
+    "json-buffer": "^3.0.1",
+    "json-schema-traverse": "^0.4.1",
+    "json-stable-stringify-without-jsonify": "^1.0.1",
+    "json5": "^2.2.3",
+    "kareem": "^2.6.3",
+    "keyv": "^4.5.4",
+    "levn": "^0.4.1",
+    "locate-path": "^6.0.0",
+    "lodash": "^4.17.21",
+    "lodash.merge": "^4.6.2",
+    "lru-cache": "^5.1.1",
+    "make-dir": "^3.1.0",
+    "math-intrinsics": "^1.1.0",
+    "media-typer": "^1.1.0",
+    "memory-pager": "^1.5.0",
+    "merge-descriptors": "^2.0.0",
+    "mime-db": "^1.54.0",
+    "mime-types": "^3.0.1",
+    "minimatch": "^3.1.2",
+    "minipass": "^5.0.0",
+    "minizlib": "^2.1.2",
+    "mkdirp": "^1.0.4",
+    "mongodb": "^6.16.0",
+    "mongodb-connection-string-url": "^3.0.2",
     "mongoose": "^8.14.0",
-    "nodemon": "^3.1.10",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0"
+    "mpath": "^0.9.0",
+    "mquery": "^5.0.0",
+    "ms": "^2.1.3",
+    "nanoid": "^3.3.11",
+    "natural-compare": "^1.4.0",
+    "negotiator": "^1.0.0",
+    "node-addon-api": "^5.1.0",
+    "node-fetch": "^2.7.0",
+    "node-releases": "^2.0.19",
+    "nopt": "^5.0.0",
+    "normalize-path": "^3.0.0",
+    "npmlog": "^5.0.1",
+    "object-assign": "^4.1.1",
+    "object-inspect": "^1.13.4",
+    "on-finished": "^2.4.1",
+    "once": "^1.4.0",
+    "optionator": "^0.9.4",
+    "p-limit": "^3.1.0",
+    "p-locate": "^5.0.0",
+    "parent-module": "^1.0.1",
+    "parseurl": "^1.3.3",
+    "path-exists": "^4.0.0",
+    "path-is-absolute": "^1.0.1",
+    "path-key": "^3.1.1",
+    "path-to-regexp": "^8.2.0",
+    "picocolors": "^1.1.1",
+    "picomatch": "^4.0.2",
+    "postcss": "^8.5.3",
+    "prelude-ls": "^1.2.1",
+    "proxy-addr": "^2.0.7",
+    "pstree.remy": "^1.1.8",
+    "punycode": "^2.3.1",
+    "qs": "^6.14.0",
+    "range-parser": "^1.2.1",
+    "raw-body": "^3.0.0",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "react-refresh": "^0.17.0",
+    "readable-stream": "^3.6.2",
+    "readdirp": "^3.6.0",
+    "require-directory": "^2.1.1",
+    "resolve-from": "^4.0.0",
+    "rimraf": "^3.0.2",
+    "rollup": "^4.40.0",
+    "router": "^2.2.0",
+    "rxjs": "^7.8.2",
+    "safe-buffer": "^5.2.1",
+    "safer-buffer": "^2.1.2",
+    "scheduler": "^0.26.0",
+    "semver": "^6.3.1",
+    "send": "^1.2.0",
+    "serve-static": "^2.2.0",
+    "set-blocking": "^2.0.0",
+    "setprototypeof": "^1.2.0",
+    "shebang-command": "^2.0.0",
+    "shebang-regex": "^3.0.0",
+    "shell-quote": "^1.8.2",
+    "side-channel": "^1.1.0",
+    "side-channel-list": "^1.0.0",
+    "side-channel-map": "^1.0.1",
+    "side-channel-weakmap": "^1.0.2",
+    "sift": "^17.1.3",
+    "signal-exit": "^3.0.7",
+    "simple-update-notifier": "^2.0.0",
+    "source-map-js": "^1.2.1",
+    "sparse-bitfield": "^3.0.3",
+    "statuses": "^2.0.1",
+    "string_decoder": "^1.3.0",
+    "string-width": "^4.2.3",
+    "strip-ansi": "^6.0.1",
+    "strip-json-comments": "^3.1.1",
+    "supports-color": "^7.2.0",
+    "tar": "^6.2.1",
+    "tinyglobby": "^0.2.13",
+    "to-regex-range": "^5.0.1",
+    "toidentifier": "^1.0.1",
+    "touch": "^3.1.1",
+    "tr46": "^5.1.1",
+    "tree-kill": "^1.2.2",
+    "tslib": "^2.8.1",
+    "type-check": "^0.4.0",
+    "type-is": "^2.0.1",
+    "undefsafe": "^2.0.5",
+    "unpipe": "^1.0.0",
+    "update-browserslist-db": "^1.1.3",
+    "uri-js": "^4.4.1",
+    "util-deprecate": "^1.0.2",
+    "vary": "^1.1.2",
+    "vite": "^6.3.2",
+    "webidl-conversions": "^7.0.0",
+    "whatwg-url": "^14.2.0",
+    "which": "^2.0.2",
+    "wide-align": "^1.1.5",
+    "word-wrap": "^1.2.5",
+    "wrap-ansi": "^7.0.0",
+    "wrappy": "^1.0.2",
+    "y18n": "^5.0.8",
+    "yallist": "^3.1.1",
+    "yargs": "^17.7.2",
+    "yargs-parser": "^21.1.1",
+    "yocto-queue": "^0.1.0"
   },
   "devDependencies": {
-    "@eslint/js": "^9.22.0",
-    "@types/react": "^19.0.10",
-    "@types/react-dom": "^19.0.4",
-    "@vitejs/plugin-react": "^4.3.4",
-    "eslint": "^9.22.0",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
-    "globals": "^16.0.0",
-    "vite": "^6.3.1"
+    "concurrently": "^9.1.2",
+    "nodemon": "^3.1.10"
   }
 }
