@@ -1,20 +1,35 @@
-const { useEffect } = require("react")
-import { useState, useEffect } from "react";
+import { useState } from 'react'
+import reactLogo from './assets/react.svg'
+import viteLogo from '/vite.svg'
+import './App.css'
 
-function App(){
-    const [message, setMessage] = useState('')
+function App() {
+  const [count, setCount] = useState(0)
 
-    useEffect(()=>{
-        fetch('http://localhost:3000/api/message')
-        .then((res) => res.json())
-        .then((data) => setMessage(data.message))
-    }, []);
-
-    return(
-        <div style={{ padding: 20}}>
-        <p>Backend message: {message}</p>
-        </div>
-    )
+  return (
+    <>
+      <div>
+        <a href="https://vite.dev" target="_blank">
+          <img src={viteLogo} className="logo" alt="Vite logo" />
+        </a>
+        <a href="https://react.dev" target="_blank">
+          <img src={reactLogo} className="logo react" alt="React logo" />
+        </a>
+      </div>
+      <h1>Vite + React</h1>
+      <div className="card">
+        <button onClick={() => setCount((count) => count + 1)}>
+          count is {count}
+        </button>
+        <p>
+          Edit <code>src/App.jsx</code> and save to test HMR
+        </p>
+      </div>
+      <p className="read-the-docs">
+        Click on the Vite and React logos to learn more
+      </p>
+    </>
+  )
 }
 
 export default App
