@@ -1,10 +1,10 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './.index.css'
+import './index.css'
 import App from './App.jsx'
 
 createRoot(document.getElementById('root')).render(
-    <StrictMode>
-         <App />
-    </StrictMode>
+  <StrictMode>
+    <App />
+  </StrictMode>,
 )
