@@ -1,11 +1,12 @@
-require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
-const path = require('path')
+const path = require('path');
+require('dotenv').config();
+
 const app = express();
 const PORT = process.env.PORT || 3000;
 
-//app.use(express.static(path.join(__dirname, '../client/dist')))
+app.use(express.static(path.join(__dirname, '../client/dist')));
 
 
 //Middleware
@@ -23,7 +24,7 @@ app.get('/api/message', (req, res)=>{
 })
 
 app.get('*', (req, res)=>{
-    res.sendFile(path.join(__dirname, '../dist/index.html'))
+    res.sendFile(path.join(__dirname, '../client/dist/index.html'))
 })
 
 app.listen(PORT, ()=>{
